+++ test/Adapter/CallbackTest.php
@@ -1,18 +1,17 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-authentication for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-authentication/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-authentication/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Authentication\Adapter;
+namespace LaminasTest\Authentication\Adapter;
 
 use Exception;
+use Laminas\Authentication\Adapter\Callback;
+use Laminas\Authentication\Result;
 use PHPUnit_Framework_TestCase as TestCase;
-use Zend\Authentication\Adapter\Callback;
-use Zend\Authentication\Result;
 
 class CallbackTest extends TestCase
 {
@@ -47,7 +46,7 @@ class CallbackTest extends TestCase
     public function testSetCallbackThrowsException()
     {
         $this->setExpectedException(
-            'Zend\Authentication\Exception\InvalidArgumentException',
+            'Laminas\Authentication\Exception\InvalidArgumentException',
             'Invalid callback provided'
         );
         $this->adapter->setCallback('This is not a valid callback');
@@ -81,7 +80,7 @@ class CallbackTest extends TestCase
     public function testAuthenticateThrowsException()
     {
         $this->setExpectedException(
-            'Zend\Authentication\Exception\RuntimeException',
+            'Laminas\Authentication\Exception\RuntimeException',
             'No callback provided'
         );
         $this->adapter->authenticate();