+++ test/Adapter/Http/ApacheResolverTest.php
@@ -1,18 +1,17 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-authentication for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-authentication/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-authentication/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Authentication\Adapter\Http;
+namespace LaminasTest\Authentication\Adapter\Http;
 
-use Zend\Authentication\Adapter\Http\ApacheResolver as Apache;
+use Laminas\Authentication\Adapter\Http\ApacheResolver as Apache;
 
 /**
- * @group      Zend_Auth
+ * @group      Laminas_Auth
  */
 class ApacheResolverTest extends \PHPUnit_Framework_TestCase
 {
@@ -75,7 +74,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetFileInvalid()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
+        $this->setExpectedException('Laminas\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
         $this->_apache->setFile($this->_badPath);
     }
 
@@ -97,7 +96,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testConstructInvalid()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
+        $this->setExpectedException('Laminas\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
         $apache = new Apache($this->_badPath);
     }
 
@@ -122,7 +121,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->_apache->setFile($file);
         $result = $this->_apache->resolve('test', null, 'password');
-        $this->assertInstanceOf('Zend\Authentication\Result', $result);
+        $this->assertInstanceOf('Laminas\Authentication\Result', $result);
         $this->assertTrue($result->isValid());
     }
 
@@ -136,7 +135,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->_apache->setFile($file);
         $result = $this->_apache->resolve('test', 'realm', 'password');
-        $this->assertInstanceOf('Zend\Authentication\Result', $result);
+        $this->assertInstanceOf('Laminas\Authentication\Result', $result);
         $this->assertTrue($result->isValid());
     }
 
@@ -149,7 +148,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->_apache->setFile($file);
         $result = $this->_apache->resolve('foo', null, 'password');
-        $this->assertInstanceOf('Zend\Authentication\Result', $result);
+        $this->assertInstanceOf('Laminas\Authentication\Result', $result);
         $this->assertFalse($result->isValid());
     }
 
@@ -162,7 +161,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->_apache->setFile($file);
         $result = $this->_apache->resolve('test', null, 'bar');
-        $this->assertInstanceOf('Zend\Authentication\Result', $result);
+        $this->assertInstanceOf('Laminas\Authentication\Result', $result);
         $this->assertFalse($result->isValid());
     }
 
@@ -173,7 +172,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
     {
         $this->_apache->setFile($this->_digest);
         $result = $this->_apache->resolve('test', 'auth', 'password');
-        $this->assertInstanceOf('Zend\Authentication\Result', $result);
+        $this->assertInstanceOf('Laminas\Authentication\Result', $result);
         $this->assertTrue($result->isValid());
     }
 }