+++ test/Adapter/Http/ProxyTest.php
@@ -1,21 +1,20 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-authentication for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-authentication/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-authentication/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Authentication\Adapter\Http;
+namespace LaminasTest\Authentication\Adapter\Http;
 
-use Zend\Authentication\Adapter\Http;
-use Zend\Http\Headers;
-use Zend\Http\Request;
-use Zend\Http\Response;
+use Laminas\Authentication\Adapter\Http;
+use Laminas\Http\Headers;
+use Laminas\Http\Request;
+use Laminas\Http\Response;
 
 /**
- * @group      Zend_Auth
+ * @group      Laminas_Auth
  */
 class ProxyTest extends \PHPUnit_Framework_TestCase
 {
@@ -135,7 +134,7 @@ class ProxyTest extends \PHPUnit_Framework_TestCase
         $digest = $this->_digestChallenge();
 
         // Make sure the result is false
-        $this->assertInstanceOf('Zend\\Authentication\\Result', $result);
+        $this->assertInstanceOf('Laminas\\Authentication\\Result', $result);
         $this->assertFalse($result->isValid());
 
         // Verify the status code and the presence of both challenges
@@ -331,7 +330,7 @@ class ProxyTest extends \PHPUnit_Framework_TestCase
         }
 
         // Create the HTTP Auth adapter
-        $a = new \Zend\Authentication\Adapter\Http($use);
+        $a = new \Laminas\Authentication\Adapter\Http($use);
         $a->setBasicResolver($this->_basicResolver);
         $a->setDigestResolver($this->_digestResolver);
 
@@ -373,8 +372,8 @@ class ProxyTest extends \PHPUnit_Framework_TestCase
     {
         $nc       = '00000001';
         $timeout  = ceil(time() / 300) * 300;
-        $nonce    = md5($timeout . ':PHPUnit:Zend\\Authentication\\Adapter\\Http');
-        $opaque   = md5('Opaque Data:Zend\\Authentication\\Adapter\\Http');
+        $nonce    = md5($timeout . ':PHPUnit:Laminas\\Authentication\\Adapter\\Http');
+        $opaque   = md5('Opaque Data:Laminas\\Authentication\\Adapter\\Http');
         $cnonce   = md5('cnonce');
         $response = md5(md5($user . ':' . $this->_digestConfig['realm'] . ':' . $pass) . ":$nonce:$nc:$cnonce:auth:"
                   . md5('GET:/'));
@@ -405,7 +404,7 @@ class ProxyTest extends \PHPUnit_Framework_TestCase
         extract($data); // $result, $status, $headers
 
         // Make sure the result is false
-        $this->assertInstanceOf('Zend\\Authentication\\Result', $result);
+        $this->assertInstanceOf('Laminas\\Authentication\\Result', $result);
         $this->assertFalse($result->isValid());
 
         // Verify the status code and the presence of the challenge
@@ -439,7 +438,7 @@ class ProxyTest extends \PHPUnit_Framework_TestCase
         extract($data); // $result, $status, $headers
 
         // Make sure the result is true
-        $this->assertInstanceOf('Zend\\Authentication\\Result', $result);
+        $this->assertInstanceOf('Laminas\\Authentication\\Result', $result);
         $this->assertTrue($result->isValid(), var_export($result->getMessages(), 1));
 
         // Verify we got a 200 response
@@ -457,7 +456,7 @@ class ProxyTest extends \PHPUnit_Framework_TestCase
         extract($data); // $result, $status, $headers
 
         // Make sure the result is false
-        $this->assertInstanceOf('Zend\\Authentication\\Result', $result);
+        $this->assertInstanceOf('Laminas\\Authentication\\Result', $result);
         $this->assertFalse($result->isValid());
 
         // Make sure it set the right HTTP code