+++ CHANGELOG.md
@@ -18,15 +18,15 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#8](https://github.com/zendframework/zend-authentication/pull/8) updates
-  dependencies to allow usage of zend-stdlib 3.0, and to require tests to
+- [zendframework/zend-authentication#8](https://github.com/zendframework/zend-authentication/pull/8) updates
+  dependencies to allow usage of laminas-stdlib 3.0, and to require tests to
   pass against PHP 7.
 
 ## 2.5.2 - 2015-06-15
 
 ### Added
 
-- [#4](https://github.com/zendframework/zend-authentication/pull/4) adds
+- [zendframework/zend-authentication#4](https://github.com/zendframework/zend-authentication/pull/4) adds
   documentation, which can be compiled using [bookdown](http://bookdown.io):
   `bookdown doc/bookdown.json`; docs can then be viewed by starting a web server
   via `php -S 0.0.0.0:8000 -t doc/html/` and browsing to http://localhost:8000/.